Implementation for ZipCodeProcessor to respond next Sabbath time

+++ SabbathText.Core/Entities/Location.cs
@@ -9,9 +9,9 @@ namespace SabbathText.Core.Entities
         public string LocationName { get; set; }
         public string Country { get; set; }
         public string Region { get; set; }
-        public decimal Longitude { get; set; }
-        public decimal Latitude { get; set; }
-        public decimal TimeZoneOffset { get; set; }
+        public double Longitude { get; set; }
+        public double Latitude { get; set; }
+        public double TimeZoneOffset { get; set; }
         public DateTime CreationTime { get; set; }
         public DateTime UpdateTime { get; set; }
     }
